feat: 가게 리뷰 평점 변경 이벤트 핸들러 구현

+++ src/main/java/com/sparta/no1delivery/domain/store/presentation/dto/StoreResponseDto.java
@@ -16,7 +16,7 @@ public record StoreResponseDto(
         String detailAddress,
         StoreStatus status,
         Double rating,
-        Integer reviewCount,
+        Long reviewCount,
         String ownerName
 ) {
     public static StoreResponseDto fromList(Store store) {
